Migrate 'drift_correct_image' and associated tests to their own modules
